requirements.txt
torch==1.10.0+cu113
torchaudio===0.10.0+cu113
torchvision==0.11.1+cu113
albumentations==1.1.0
hydra-core==1.2.0

matplotlib==3.2.2
pandas==1.3.1
torchinfo==1.7.1
tqdm==4.64.1
wandb==0.12.6
